third_party/blink/web_tests/plugins/mouse-capture-inside-shadow-expected.txt

CONSOLE WARNING: line 23: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
CONSOLE MESSAGE: Blink Test Plugin: MouseDown at (12,12)
CONSOLE MESSAGE: Blink Test Plugin: MouseUp at (212,112)
This tests that mousedown captures the mouse correctly for a plugin object and continues to send events even if the mouse moves outside of the bounds of the plugin. On a successful run, plugin will see mousedown, followed by a mousemove, followed by a mouseup.
Console
plugin.mousedown
plugin.mousemove
plugin.mouseup
